initial port to nalgebra

+++ libs/genetic-algorithm/src/crossover/uniform.rs
@@ -32,8 +32,8 @@ mod test {
     #[test]
     fn test_crossover() {
         let mut rng = ChaCha8Rng::from_seed(Default::default());
-        let parent_a: Chromosome = (1..=100).map(|x| x as f32).collect();
-        let parent_b: Chromosome = (1..=100).map(|x| -x as f32).collect();
+        let parent_a: Chromosome = (1..=100).map(|x| x as f64).collect();
+        let parent_b: Chromosome = (1..=100).map(|x| -x as f64).collect();
         let child = UniformCrossover::default().crossover(&mut rng, &parent_a, &parent_b);
         let diff_a = child.iter().zip(parent_a).filter(|(c, p)| *c != p).count();
         let diff_b = child.iter().zip(parent_b).filter(|(c, p)| *c != p).count();
